Switches monetary balances to BigDecimal for accuracy
Replaces floating-point types with BigDecimal for all account
balance fields, calculations, and mappings to prevent rounding
errors and ensure precision in monetary operations.

Updates test cases and mapping logic accordingly to maintain
consistency throughout the service and data layers.

+++ banking/src/main/java/es/triana/company/banking/service/TransactionService.java
@@ -231,11 +231,11 @@ public class TransactionService {
         }
 
         if (account.getLastBalanceReal() != null) {
-            return account.getLastBalanceReal();
+            return account.getLastBalanceReal().doubleValue();
         }
 
         if (account.getLastBalanceAvailable() != null) {
-            return account.getLastBalanceAvailable();
+            return account.getLastBalanceAvailable().doubleValue();
         }
 
         return 0.0;
@@ -404,11 +404,11 @@ public class TransactionService {
 
     private double resolveCurrentBalance(Account account) {
         if (account.getLastBalanceReal() != null) {
-            return account.getLastBalanceReal();
+            return account.getLastBalanceReal().doubleValue();
         }
 
         if (account.getLastBalanceAvailable() != null) {
-            return account.getLastBalanceAvailable();
+            return account.getLastBalanceAvailable().doubleValue();
         }
 
         return 0.0;
